+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 10.6
 ---------------
 - Add Pashto to right-to-left languages in Warehouse.php, Inputs.php & PDF.php
+- CSS FixedMenu bug when menu hidden in stylesheet.css
 
 Changes in 10.5.2
 -----------------