+++ CHANGES.md
@@ -14,6 +14,7 @@ Changes in 10.6
 - CSS responsive Fix list overflow-x scroll inside div.st in stylesheet.css
 - CSS set max-width for City & State select in Address.inc.php
 - Security fix for dynamic include in index.php
+- Fix SQL limit 1 when adding existing Contact in Address.inc.php
 
 Changes in 10.5.2
 -----------------